Added Node superclass, cleaned up code
git-svn-id: https://subversion.assembla.com/svn/elec476/trunk/src@27 4db04919-9836-40c1-aee8-a115cec00ef4

+++ contextAwareRouting/Mainline.java
@@ -33,7 +33,7 @@ public class Mainline {
 		 Server = new CentralServer(Users, Relays);
 		 
 		 //Create random arrival times
-		 ArrayList<Integer> Arrival = Rand.poissonList(requestrate, maxtime);
+		 ArrayList<Double> Arrival = Rand.poissonList(requestrate, maxtime);
 		 
 		 //time setup
 		 boolean done = false;
